[Studio UI] Fall back to main-widget context when active tabset is not restored (#2716)
Extract getNextTabId into its own util and make getOpenedMainWidget fall
back to the authoritative main-widget context, since Model.fromJson does
not restore the active tabset after closing the active pane of a split.
